chore(names): rename ResCardOps and DevCardOps in ResourceCardOps and DevelopmentCardOps
