+++ src/test/scala/scatan/model/ScatanEffectsTest.scala
@@ -4,7 +4,7 @@ import scatan.BaseTest
 import scatan.model.components.{ResourceCard, ResourceType}
 import scatan.model.game.ScatanEffects.{NextTurnEffect, PlaceRobberEffect, RollEffect, TradeWithPlayerEffect}
 import scatan.model.game.config.ScatanPlayer
-import scatan.model.game.ops.ResCardOps.assignResourceCard
+import scatan.model.game.ops.ResourceCardOps.assignResourceCard
 import scatan.model.game.{ScatanGame, ScatanState}
 
 class ScatanEffectsTest extends BaseTest:
