+++ src/test/scala/scatan/model/game/ops/BuildingOpsTest.scala
@@ -4,7 +4,7 @@ import scatan.model.components.*
 import scatan.model.game.config.ScatanPlayer
 import scatan.model.game.ops.BuildingOps.{assignBuilding, build}
 import scatan.model.game.ops.EmptySpotOps.{emptyRoadSpot, emptyStructureSpot}
-import scatan.model.game.ops.ResCardOps.assignResourceCard
+import scatan.model.game.ops.ResourceCardOps.assignResourceCard
 import scatan.model.game.{BaseScatanStateTest, ScatanState}
 import scatan.model.map.{RoadSpot, StructureSpot}
