+++ src/test/scala/scatan/model/game/ops/DevelopmentCardOpsTest.scala
@@ -2,11 +2,11 @@ package scatan.model.game.ops
 
 import scatan.lib.game.Game
 import scatan.model.components.{DevelopmentCard, DevelopmentType, ResourceCard, ResourceType}
-import scatan.model.game.ops.DevCardOps.{assignDevelopmentCard, buyDevelopmentCard, consumeDevelopmentCard}
-import scatan.model.game.ops.ResCardOps.assignResourceCard
+import scatan.model.game.ops.DevelopmentCardOps.{assignDevelopmentCard, buyDevelopmentCard, consumeDevelopmentCard}
+import scatan.model.game.ops.ResourceCardOps.assignResourceCard
 import scatan.model.game.{BaseScatanStateTest, ScatanState}
 
-class DevCardOpsTest extends BaseScatanStateTest:
+class DevelopmentCardOpsTest extends BaseScatanStateTest:
 
   "A State with development cards Ops" should "have empty development cards when game start" in {
     val state = ScatanState(threePlayers)
