+++ src/test/scala/scatan/model/game/ops/ResourceCardOpsTest.scala
@@ -3,13 +3,13 @@ package scatan.model.game.ops
 import scatan.model.components.{BuildingType, ResourceCard, ResourceCards, ResourceType}
 import scatan.model.game.ops.BuildingOps.assignBuilding
 import scatan.model.game.ops.EmptySpotOps.emptyStructureSpot
-import scatan.model.game.ops.ResCardOps.*
+import scatan.model.game.ops.ResourceCardOps.*
 import scatan.model.game.{BaseScatanStateTest, ScatanState}
 import scatan.model.map.HexagonInMap.layer
 import scatan.model.map.{RoadSpot, Spot, StructureSpot}
 import scatan.utils.UnorderedTriple
 
-class ResCardOpsTest extends BaseScatanStateTest:
+class ResourceCardOpsTest extends BaseScatanStateTest:
 
   extension (state: ScatanState)
     /** This method assigns resources to players based on the number of the hexagons where their buildings are located.
